requirements.txt
torch==2.0.0
torchaudio==2.0.0
speechbrain==0.5.16
librosa==0.10.0
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0
soundfile==0.12.1
pyyaml==6.0
pydantic==2.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorboard==2.14.1
matplotlib==3.8.0
tqdm==4.66.1
pytest==7.4.2
pytest-cov==4.1.0
